+++ src/plugin.ts
@@ -7,6 +7,7 @@ import { getDirname } from 'cross-dirname';
 import { noCycle } from './rules/cycle/cycle.js';
 import { noEntryPointImports } from './rules/entryPoint/entryPoint.js';
 import { noExternalBarrelReexports } from './rules/externalBarrelReexports/externalBarrelReexports.js';
+import { namedAsDefault } from './rules/namedAsDefault/namedAsDefault.js';
 import { nodePrefix } from './rules/nodePrefix/nodePrefix.js';
 import { noTestImportsInProd } from './rules/testInProd/testInProd.js';
 import { noUnresolvedImports } from './rules/unresolved/unresolved.js';
@@ -38,6 +39,7 @@ const plugin = {
     'no-unresolved-imports': noUnresolvedImports,
     'no-external-barrel-reexports': noExternalBarrelReexports,
     'no-test-imports-in-prod': noTestImportsInProd,
+    'no-named-as-default': namedAsDefault,
     'require-node-prefix': nodePrefix,
   },
   processors: {},
@@ -54,6 +56,7 @@ const recommendedConfig = {
     'fast-import/no-unresolved-imports': 'error',
     'fast-import/no-external-barrel-reexports': 'error',
     'fast-import/no-test-imports-in-prod': 'error',
+    'fast-import/no-named-as-default': 'error',
     'fast-import/require-node-prefix': 'off',
   },
 } as const;
@@ -69,6 +72,7 @@ const allConfig = {
     'fast-import/no-unresolved-imports': 'error',
     'fast-import/no-external-barrel-reexports': 'error',
     'fast-import/no-test-imports-in-prod': 'error',
+    'fast-import/no-named-as-default': 'error',
     'fast-import/require-node-prefix': 'error',
   },
 } as const;
@@ -84,6 +88,7 @@ const offConfig = {
     'fast-import/no-unresolved-imports': 'off',
     'fast-import/no-external-barrel-reexports': 'off',
     'fast-import/no-test-imports-in-prod': 'off',
+    'fast-import/no-named-as-default': 'off',
     'fast-import/require-node-prefix': 'off',
   },
 } as const;