+++ src/rules/namedAsDefault/__test__/project/b.ts
@@ -1,0 +1,2 @@
+export const b = 10;
+export default 20;